requirements.txt

torch>=2.2.0
torchvision>=0.17.0
timm>=1.0.9
gradio>=4.44.0
pandas>=2.1.4
scikit-learn>=1.4.0
numpy>=1.26.4
Pillow>=10.2.0
uvicorn>=0.30.0
fastapi>=0.115.0
